requirements.txt
flask
flask-smorest
python-dotenv
sqlalchemy
flask-sqlalchemy
flask-migrate
psycopg2-binary
pydantic
pydantic-settings
asgiref
